# server/docs/swagger.yaml
definitions:
...
  request.BudgetRuleOperateParams:
    properties:
      budgetDuration:
        description: 预算周期(1d/7d/30d)
        type: string
      budgetHardLimit:
        description: 硬限
        type: boolean
      budgetLimit:
        description: 预算上限(¥,0=不限)
        type: number
      isActive:
        description: 启停(nil=不改)
        type: boolean
      ruleId:
        description: 规则ID(修改时必填)
        example: "0"
        type: string
      scopeId:
        description: 对象ID(部门ID/用户ID)
        example: "0"
        type: string
      scopeType:
        description: 维度(dept/user)
        type: string
      softWarnPercent:
        description: 软限预警阈值(%)
        type: integer
    type: object
...
  request.IdsReq:
    properties:
      ids:
        description: ID列表(string对齐前端,雪花ID传输为string)
        items:
          type: string
        type: array
    type: object
...
  request.SettingConfig:
    properties:
      auth:
        $ref: '#/definitions/system.SysAuthConfig'
      general:
        $ref: '#/definitions/system.SysGeneralConfig'
      ldap:
        $ref: '#/definitions/system.SysLdapConfig'
      notify:
        $ref: '#/definitions/system.SysNotifyConfig'
      notifyPolicy:
        $ref: '#/definitions/system.SysNotifyPolicy'
      security:
        $ref: '#/definitions/system.SysSecurityConfig'
    type: object
...
  request.TestEmailReq:
    properties:
      emailFromAddr:
        type: string
      emailFromName:
        type: string
      emailHost:
        type: string
      emailPassword:
        type: string
      emailPort:
        type: integer
      emailSSLMode:
        type: string
      emailUsername:
        type: string
      testTo:
        type: string
    type: object
  request.TestWecomAppReq:
    properties:
      redirectBase:
        description: 当前表单跳转base(未保存也可测;空降级纯文本)
        type: string
      testUserId:
        description: 目标用户(须已绑定企微)
        example: "0"
        type: string
    required:
    - testUserId
    type: object
  request.TestWecomBotReq:
    properties:
      webhookUrl:
        description: 当前表单 webhook(未保存也可测)
        type: string
    required:
    - webhookUrl
    type: object
...
  response.BudgetRuleView:
    properties:
      budgetDuration:
        description: 预算周期
        type: string
      budgetHardLimit:
        description: 硬限
        type: boolean
      budgetLimit:
        description: 预算上限(¥,0=不限)
        type: number
      budgetUsed:
        description: 已用(¥,读时聚合)
        type: number
      budgetUsedPercent:
        description: 执行率(%,限>0)
        type: number
      isActive:
        description: 是否启用
        type: boolean
      isHardLimited:
        description: 是否已触发硬限超限(本周期)
        type: boolean
      isSoftWarn:
        description: 是否已触发软限预警(本周期)
        type: boolean
      ruleId:
        description: 规则ID
        example: "0"
        type: string
      scopeId:
        description: 对象ID
        example: "0"
        type: string
      scopeName:
        description: 对象名(部门名/用户名)
        type: string
      scopeType:
        description: 维度
        type: string
      softWarnPercent:
        description: 软限阈值(%)
        type: integer
    type: object
...
  response.PageResult:
    properties:
      pageNum:
        type: integer
      pageSize:
        type: integer
      rows: {}
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
    type: object
...
  system.SysAuthConfig:
    properties:
      createTime:
        description: |-
          字段名用 gorm 约定的 CreatedAt/UpdatedAt, 由 gorm 自动维护(写入填值、更新刷值);
          column 锁定 create_time/update_time 列名, 沿用历史列不漂移成 created_at/updated_at;
          json 仍叫 createTime/updateTime 对齐前端 CommonRecord。
        type: string
      giteeCallbackUrl:
        type: string
      giteeClientId:
        type: string
      giteeClientSecret:
        type: string
      giteeEnabled:
        description: Gitee
        type: boolean
      githubCallbackUrl:
        type: string
      githubClientId:
        type: string
      githubClientSecret:
        type: string
      githubEnabled:
        description: GitHub
        type: boolean
      id:
        description: 主键(过渡保留,未改造系统表用)
        example: "0"
        type: string
      registerEnabled:
        description: 账号功能
        type: boolean
      resetPwdEnabled:
        type: boolean
      updateTime:
        type: string
      wechatCallbackUrl:
        type: string
      wechatClientId:
        type: string
      wechatClientSecret:
        type: string
      wechatEnabled:
        description: 微信开放平台
        type: boolean
      wecomAgentId:
        type: integer
      wecomCallbackUrl:
        type: string
      wecomClientId:
        type: string
      wecomClientSecret:
        type: string
      wecomCorpId:
        type: string
      wecomDomainFileContent:
        type: string
      wecomDomainFileName:
        type: string
      wecomEnabled:
        description: 企业微信
        type: boolean
    type: object
...
  system.SysGeneralConfig:
    properties:
      createTime:
        description: |-
          字段名用 gorm 约定的 CreatedAt/UpdatedAt, 由 gorm 自动维护(写入填值、更新刷值);
          column 锁定 create_time/update_time 列名, 沿用历史列不漂移成 created_at/updated_at;
          json 仍叫 createTime/updateTime 对齐前端 CommonRecord。
        type: string
      defaultPassword:
        description: 导入新建/复活用户的初始密码;首登强制改密
        type: string
      defaultRoleId:
        example: "0"
        type: string
      faviconUrl:
        type: string
      id:
        description: 主键(过渡保留,未改造系统表用)
        example: "0"
        type: string
      loginLogRetentionDays:
        type: integer
      logoUrl:
        type: string
      operationLogRetentionDays:
        type: integer
      systemDescription:
        type: string
      systemName:
        type: string
      updateTime:
        type: string
    type: object
  system.SysLdapConfig:
    properties:
      attrEmail:
        type: string
      attrNickname:
        type: string
      attrUsername:
        description: 属性映射
        type: string
      autoCreate:
        description: 用户策略
        type: boolean
      baseDN:
        type: string
      bindDN:
        type: string
      bindPass:
        type: string
      createTime:
        description: |-
          字段名用 gorm 约定的 CreatedAt/UpdatedAt, 由 gorm 自动维护(写入填值、更新刷值);
          column 锁定 create_time/update_time 列名, 沿用历史列不漂移成 created_at/updated_at;
          json 仍叫 createTime/updateTime 对齐前端 CommonRecord。
        type: string
      enabled:
        description: 连接配置
        type: boolean
      filter:
        type: string
      host:
        type: string
      id:
        description: 主键(过渡保留,未改造系统表用)
        example: "0"
        type: string
      port:
        type: integer
      updateTime:
        type: string
      useSSL:
        type: boolean
    type: object
...
  system.SysNotifyConfig:
    properties:
      createTime:
        description: |-
          字段名用 gorm 约定的 CreatedAt/UpdatedAt, 由 gorm 自动维护(写入填值、更新刷值);
          column 锁定 create_time/update_time 列名, 沿用历史列不漂移成 created_at/updated_at;
          json 仍叫 createTime/updateTime 对齐前端 CommonRecord。
        type: string
      emailEnabled:
        description: 邮件通知
        type: boolean
      emailFromAddr:
        type: string
      emailFromName:
        type: string
      emailHost:
        type: string
      emailPassword:
        type: string
      emailPort:
        type: integer
      emailSSLMode:
        description: none / ssl / starttls
        type: string
      emailUsername:
        type: string
      id:
        description: 主键(过渡保留,未改造系统表用)
        example: "0"
        type: string
      pushBudgetAlertEnabled:
        type: boolean
      pushMorningReportEnabled:
        description: 事件开关(控制该事件是否走外部渠道;站内通知不受控)
        type: boolean
      updateTime:
        type: string
      webhookEnabled:
        description: Webhook
        type: boolean
      webhookSecret:
        type: string
      webhookUrl:
        type: string
      wecomBotEnabled:
        description: 企业微信群机器人(webhook,markdown 进群)
        type: boolean
      wecomBotWebhook:
        type: string
      wecomPushEnabled:
        description: 企业微信应用消息(凭证复用 sys_auth_config 企微段,此处不重复存)
        type: boolean
      wecomPushMaxTargets:
        type: integer
      wecomPushRedirectBase:
        type: string
    type: object
  system.SysNotifyPolicy:
    properties:
      createTime:
        description: |-
          字段名用 gorm 约定的 CreatedAt/UpdatedAt, 由 gorm 自动维护(写入填值、更新刷值);
          column 锁定 create_time/update_time 列名, 沿用历史列不漂移成 created_at/updated_at;
          json 仍叫 createTime/updateTime 对齐前端 CommonRecord。
        type: string
      enabled:
        type: boolean
      id:
        description: 主键(过渡保留,未改造系统表用)
        example: "0"
        type: string
      params:
        items:
          type: integer
        type: array
      sceneKey:
        type: string
      sendTime:
        type: string
      targetIds:
        items:
          type: integer
        type: array
      targetType:
        type: string
      updateTime:
        type: string
    type: object
...
  system.SysSecurityConfig:
    properties:
      captchaEnabled:
        description: 验证码(登录链路验证码生成用,前端「安全配置-验证码」tab 直接消费本段)
        type: boolean
      captchaOpen:
        type: integer
      captchaTimeout:
        type: integer
      captchaTolerance:
        type: integer
      captchaType:
        type: string
      createTime:
        description: |-
          字段名用 gorm 约定的 CreatedAt/UpdatedAt, 由 gorm 自动维护(写入填值、更新刷值);
          column 锁定 create_time/update_time 列名, 沿用历史列不漂移成 created_at/updated_at;
          json 仍叫 createTime/updateTime 对齐前端 CommonRecord。
        type: string
      id:
        description: 主键(过渡保留,未改造系统表用)
        example: "0"
        type: string
      imgHeight:
        type: integer
      imgWidth:
        type: integer
      ipBlacklist:
        type: string
      ipValidationEnabled:
        description: IP 校验(对齐前端 SecuritySettingConfig.ipValidation*)
        type: boolean
      ipValidationMode:
        type: string
      ipWhitelist:
        type: string
      keyLong:
        type: integer
      limitCount:
        type: integer
      limitEnable:
        description: 限流(后端登录链路用,前端不直接消费)
        type: boolean
      limitWindow:
        type: integer
      loginFailLockCount:
        description: 登录失败锁定(对齐前端 SecuritySettingConfig.loginFailLock*)
        type: integer
      loginFailLockTime:
        type: integer
      passwordMinLength:
        description: 密码复杂度(对齐前端 SecuritySettingConfig.password*)
        type: integer
      passwordRequireDigit:
        type: boolean
      passwordRequireLowercase:
        type: boolean
      passwordRequireSpecial:
        type: boolean
      passwordRequireUppercase:
        type: boolean
      pwdExpireDays:
        type: integer
      pwdExpireEnable:
        description: 密码过期(后端登录链路用,前端不直接消费)
        type: boolean
      updateTime:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用devops-admin进行开发基础平台
  title: Devops-Admin Swagger API接口文档
  version: v0.2.0
paths:
...
  /gateway/budget:
    delete:
      parameters:
      - description: 规则ID列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 批量删除预算规则
      tags:
      - GatewayBudget
    post:
      parameters:
      - description: 规则配置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BudgetRuleOperateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BudgetRuleView'
                msg:
                  type: string
              type: object
      summary: 新增预算规则
      tags:
      - GatewayBudget
    put:
      parameters:
      - description: 规则配置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BudgetRuleOperateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BudgetRuleView'
                msg:
                  type: string
              type: object
      summary: 修改预算规则
      tags:
      - GatewayBudget
...
  /gateway/budget/list:
    get:
      parameters:
      - description: 维度(dept/user,空=全部)
        in: query
        name: scopeType
        type: string
      - description: 启停(1/0,nil=全部)
        in: query
        name: isActive
        type: integer
      - description: 页码
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      rows:
                        items:
                          $ref: '#/definitions/response.BudgetRuleView'
                        type: array
                    type: object
                msg:
                  type: string
              type: object
      summary: 分页获取预算规则(含读时聚合已用+预警状态)
      tags:
      - GatewayBudget
  /gateway/budget/summary:
    get:
      parameters:
      - description: 维度(all/self,超管默认all)
        in: query
        name: scope
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      summary: 预算汇总(按维度:Key/部门/用户)
      tags:
      - GatewayBudget
...
  /system/setting:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/request.SettingConfig'
                msg:
                  type: string
              type: object
      summary: 获取系统设置(管理员)
      tags:
      - SysSetting
    put:
      consumes:
      - application/json
      parameters:
      - description: 系统设置(general/security/ldap/notify/auth 任一可选)
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SettingConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: boolean
                msg:
                  type: string
              type: object
      summary: 更新系统设置(管理员)
      tags:
      - SysSetting
  /system/setting/notify/test-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮件配置+收件人地址
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.TestEmailReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: boolean
                msg:
                  type: string
              type: object
      summary: 发送测试邮件(使用当前表单值,无需先保存)
      tags:
      - SysSetting
  /system/setting/notify/test-wecom-app:
    post:
      consumes:
      - application/json
      parameters:
      - description: 目标用户+跳转base
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.TestWecomAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: boolean
                msg:
                  type: string
              type: object
      summary: 发送企微应用消息测试(用已保存企微凭证,表单跳转地址未保存也可测)
      tags:
      - SysSetting
  /system/setting/notify/test-wecom-bot:
    post:
      consumes:
      - application/json
      parameters:
      - description: 群机器人webhook地址
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.TestWecomBotReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: boolean
                msg:
                  type: string
              type: object
      summary: 发送企微群机器人测试(使用当前表单webhook,无需先保存)
      tags:
      - SysSetting
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...